+++ native/src/main/java/io/ballerina/lib/redis/ConnectionCommands.java
@@ -70,13 +70,12 @@ public class ConnectionCommands {
      * Remove all keys from all databases.
      *
      * @param redisClient Client from the Ballerina redis client
-     * @return A string with the value `OK` if the operation was successful
      */
     public static Object flushAll(BObject redisClient) {
         try {
             RedisConnectionCommandExecutor executor = getConnection(redisClient).getConnectionCommandExecutor();
-            String response = executor.flushAll();
-            return StringUtils.fromString(response);
+            executor.flushAll();
+            return null;
         } catch (Throwable e) {
             return createBError(e);
         }
@@ -86,13 +85,12 @@ public class ConnectionCommands {
      * Remove all keys from the currently selected database.
      *
      * @param redisClient Client from the Ballerina redis client
-     * @return A string with the value `OK` if the operation was successful
      */
     public static Object flushDb(BObject redisClient) {
         try {
             RedisConnectionCommandExecutor executor = getConnection(redisClient).getConnectionCommandExecutor();
-            String response = executor.flushDb();
-            return StringUtils.fromString(response);
+            executor.flushDb();
+            return null;
         } catch (Throwable e) {
             return createBError(e);
         }